Flatten source tree
The source tree was not getting along with my static analyzers (clang
and cppcheck) and just really isn't necessary for such a small project.
I was reading the bash source last night and their tree was completely
flat, as is the tree for other smaller projects like less, so I figure
a flat source tree is fine for this as well.